test(api-core): move retained boot interface tests to api-db
